Simple Solution
added a larger threshold to collisionWillOccur() this fixes most
occurances of sticky walls. Sprite still seems to go over the the wall
in certain scenarios which leads me to believe that the body position,
sprite positions and janitor position are falling out of sync somewhere.
This is more of a quick and dirty solution.