Pushing game music
Also the volumes of all sounds are editable in the Constants class.

There is a little flare of guitar that goes off towards the end of the
loop that continues into the front end of the loop, but it is definitely
quieter in the front of the loop. The game music was coming in a little
suddenly without me quiting it down. I'm willing to work on the mix a
little more if it is too noticable.